+++ docs/source/changes.md
@@ -5,10 +5,13 @@ chronological order. Releases follow [semantic versioning](https://semver.org/) 
 releases are available on [PyPI](https://pypi.org/project/pytask) and
 [Anaconda.org](https://anaconda.org/conda-forge/pytask).
 
-## 0.4.6
+## 0.5.0 - 2024-xx-xx
 
 - {pull}`548` fixes the type hints for {meth}`~pytask.Task.execute` and
   {meth}`~pytask.TaskWithoutPath.execute`. Thanks to {user}`Ostheer`.
+- {pull}`551` removes the deprecated `@pytask.mark.depends_on` and
+  `@pytask.mark.produces`.
+- {pull}`552` removes the deprecated `@pytask.mark.task`.
 
 ## 0.4.5 - 2024-01-09